♻️ (TfsWorkItemMigrationProcessorOptions): move validation logic to a separate validator class
The validation logic is moved from `TfsWorkItemMigrationProcessorOptions` to a new class `TfsWorkItemMigrationProcessorOptionsValidator`. This change improves the single responsibility principle by separating the concerns of option definition and validation. It also enhances code maintainability and readability by isolating validation logic in its own class.
