+++ src/MigrationTools.Clients.TfsObjectModel/Processors/TfsWorkItemMigrationProcessorOptions.cs
@@ -11,7 +11,7 @@ using System.DirectoryServices.AccountManagement;
 namespace MigrationTools.Processors
 {
 
-    public class TfsWorkItemMigrationProcessorOptions : ProcessorOptions, IWorkItemProcessorConfig, IValidateOptions<TfsWorkItemMigrationProcessorOptions>
+    public class TfsWorkItemMigrationProcessorOptions : ProcessorOptions, IWorkItemProcessorConfig
     {
 
 
@@ -102,59 +102,7 @@ namespace MigrationTools.Processors
         /// </summary>
         public bool SkipRevisionWithInvalidAreaPath { get; set; }
 
-        public ValidateOptionsResult Validate(string name, TfsWorkItemMigrationProcessorOptions options)
-        {
-            var errors = new List<string>();
-            ValidateOptionsResult baseResult = base.Validate(name, options);
-            if (baseResult != ValidateOptionsResult.Success)
-            {
-                errors.AddRange(baseResult.Failures);
-            }
-            // Check if WIQLQuery is provided
-            if (string.IsNullOrWhiteSpace(options.WIQLQuery))
-            {
-                errors.Add($"The WIQLQuery on {name} must be provided.");
-            }
-            else
-            {
-                // Validate the presence of required elements in the WIQL query
-                if (!ContainsTeamProjectCondition(options.WIQLQuery))
-                {
-                    errors.Add("The WIQLQuery on {name} must contain the condition '[System.TeamProject] = @TeamProject'.");
-                }
-
-                if (!UsesWorkItemsTable(options.WIQLQuery))
-                {
-                    errors.Add("The WIQLQuery on {name} must use 'WorkItems' after the 'FROM' clause and not 'WorkItemLinks'.");
-                }
-            }
-            if (errors.Count > 0)
-            {
-                return ValidateOptionsResult.Fail(errors);
-            }
-
-            return ValidateOptionsResult.Success;
-        }
-
-        private bool ContainsTeamProjectCondition(string query)
-        {
-            // Regex to match '[System.TeamProject] = @TeamProject'
-            string teamProjectPattern = @"\[System\.TeamProject\]\s*=\s*@TeamProject";
-
-            return Regex.IsMatch(query, teamProjectPattern, RegexOptions.IgnoreCase);
-        }
-
-        // Check if the WIQL query is using 'WorkItems' and not 'WorkItemLinks'
-        private bool UsesWorkItemsTable(string query)
-        {
-            // Regex to ensure that 'FROM WorkItems' exists and 'WorkItemLinks' does not follow the FROM clause
-            string fromWorkItemsPattern = @"FROM\s+WorkItems\b";
-            string fromWorkItemLinksPattern = @"FROM\s+WorkItemLinks\b";
-
-            // Return true if "WorkItems" is used and "WorkItemLinks" is not
-            return Regex.IsMatch(query, fromWorkItemsPattern, RegexOptions.IgnoreCase) &&
-                   !Regex.IsMatch(query, fromWorkItemLinksPattern, RegexOptions.IgnoreCase);
-        }
+        
     }
